Abstract PsqlReplicationCheck
